refactor(cli): Merge hooks subcommand into init
